Dynamic Client Registration
Update the SDK to support the new programmatic registration endpoints. Also add tests to ensure clients can register with these endpoints.

Update the rest of the integration tests to use dynamically-created clients for both writing and sharing. Two tests leverage email-based lookups of client information and will not work with the dynamic clients until email-based lookups are implemented for that style of client.

Fixes E3DB-661